[feat:#170][tsdb]: add series reader to support series.Filter

+++ query/series_search_test.go
@@ -17,7 +17,7 @@ func TestSampleCondition(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 	mockFilter := series.NewMockFilter(ctrl)
-	s := mockSeriesIDSet(int64(1), roaring.BitmapOf(1, 2, 3, 4))
+	s := mockSeriesIDSet(uint32(1), roaring.BitmapOf(1, 2, 3, 4))
 
 	query, _ := sql.Parse("select f from cpu")
 	search := newSeriesSearch(1, mockFilter, query)
@@ -75,20 +75,20 @@ func TestNotCondition(t *testing.T) {
 	query, _ := sql.Parse("select f from cpu where ip!='1.1.1.1'")
 	mockFilter.EXPECT().
 		FindSeriesIDsByExpr(uint32(1), &stmt.EqualsExpr{Key: "ip", Value: "1.1.1.1"}, query.TimeRange).
-		Return(mockSeriesIDSet(int64(11), roaring.BitmapOf(3, 4)), nil)
+		Return(mockSeriesIDSet(uint32(11), roaring.BitmapOf(3, 4)), nil)
 
 	mockFilter.EXPECT().
 		GetSeriesIDsForTag(uint32(1), "ip", query.TimeRange).
-		Return(mockSeriesIDSet(int64(11), roaring.BitmapOf(1, 2, 3, 4)), nil)
+		Return(mockSeriesIDSet(uint32(11), roaring.BitmapOf(1, 2, 3, 4)), nil)
 	search := newSeriesSearch(1, mockFilter, query)
 	resultSet, _ := search.Search()
-	assert.Equal(t, *mockSeriesIDSet(int64(11), roaring.BitmapOf(1, 2)), *resultSet)
+	assert.Equal(t, *mockSeriesIDSet(uint32(11), roaring.BitmapOf(1, 2)), *resultSet)
 
 	// error
 	query, _ = sql.Parse("select f from cpu where ip!='1.1.1.1'")
 	mockFilter.EXPECT().
 		FindSeriesIDsByExpr(uint32(1), &stmt.EqualsExpr{Key: "ip", Value: "1.1.1.1"}, query.TimeRange).
-		Return(mockSeriesIDSet(int64(11), roaring.BitmapOf(3, 4)), nil)
+		Return(mockSeriesIDSet(uint32(11), roaring.BitmapOf(3, 4)), nil)
 
 	mockFilter.EXPECT().
 		GetSeriesIDsForTag(uint32(1), "ip", query.TimeRange).
@@ -109,13 +109,13 @@ func TestBinaryCondition(t *testing.T) {
 		"where ip='1.1.1.1' and path='/data' and time>'20190410 00:00:00' and time<'20190410 10:00:00'")
 	mockFilter.EXPECT().
 		FindSeriesIDsByExpr(uint32(1), &stmt.EqualsExpr{Key: "ip", Value: "1.1.1.1"}, query.TimeRange).
-		Return(mockSeriesIDSet(int64(11), roaring.BitmapOf(1, 2, 3, 4)), nil)
+		Return(mockSeriesIDSet(uint32(11), roaring.BitmapOf(1, 2, 3, 4)), nil)
 	mockFilter.EXPECT().
 		FindSeriesIDsByExpr(uint32(1), &stmt.EqualsExpr{Key: "path", Value: "/data"}, query.TimeRange).
-		Return(mockSeriesIDSet(int64(11), roaring.BitmapOf(3, 5)), nil)
+		Return(mockSeriesIDSet(uint32(11), roaring.BitmapOf(3, 5)), nil)
 	search := newSeriesSearch(1, mockFilter, query)
 	resultSet, _ := search.Search()
-	assert.Equal(t, *mockSeriesIDSet(int64(11), roaring.BitmapOf(3)), *resultSet)
+	assert.Equal(t, *mockSeriesIDSet(uint32(11), roaring.BitmapOf(3)), *resultSet)
 
 	// or
 	mockFilter2 := series.NewMockFilter(ctrl)
@@ -123,13 +123,13 @@ func TestBinaryCondition(t *testing.T) {
 		"where ip='1.1.1.1' or path='/data' and time>'20190410 00:00:00' and time<'20190410 10:00:00'")
 	mockFilter2.EXPECT().
 		FindSeriesIDsByExpr(uint32(1), &stmt.EqualsExpr{Key: "ip", Value: "1.1.1.1"}, query.TimeRange).
-		Return(mockSeriesIDSet(int64(11), roaring.BitmapOf(1, 2, 3, 4)), nil)
+		Return(mockSeriesIDSet(uint32(11), roaring.BitmapOf(1, 2, 3, 4)), nil)
 	mockFilter2.EXPECT().
 		FindSeriesIDsByExpr(uint32(1), &stmt.EqualsExpr{Key: "path", Value: "/data"}, query.TimeRange).
-		Return(mockSeriesIDSet(int64(11), roaring.BitmapOf(3, 5)), nil)
+		Return(mockSeriesIDSet(uint32(11), roaring.BitmapOf(3, 5)), nil)
 	search = newSeriesSearch(1, mockFilter2, query)
 	resultSet, _ = search.Search()
-	assert.Equal(t, *mockSeriesIDSet(int64(11), roaring.BitmapOf(1, 2, 3, 4, 5)), *resultSet)
+	assert.Equal(t, *mockSeriesIDSet(uint32(11), roaring.BitmapOf(1, 2, 3, 4, 5)), *resultSet)
 
 	// error
 	mockFilter3 := series.NewMockFilter(ctrl)
@@ -146,7 +146,7 @@ func TestBinaryCondition(t *testing.T) {
 		"where ip='1.1.1.1' or path='/data' and time>'20190410 00:00:00' and time<'20190410 10:00:00'")
 	mockFilter4.EXPECT().
 		FindSeriesIDsByExpr(uint32(1), &stmt.EqualsExpr{Key: "ip", Value: "1.1.1.1"}, query.TimeRange).
-		Return(mockSeriesIDSet(int64(11), roaring.BitmapOf(1, 2, 3, 4)), nil)
+		Return(mockSeriesIDSet(uint32(11), roaring.BitmapOf(1, 2, 3, 4)), nil)
 	mockFilter4.EXPECT().
 		FindSeriesIDsByExpr(uint32(1), &stmt.EqualsExpr{Key: "path", Value: "/data"}, query.TimeRange).
 		Return(nil, errors.New("right error"))
@@ -165,35 +165,35 @@ func TestComplexCondition(t *testing.T) {
 		" where (ip not in ('1.1.1.1','2.2.2.2') and region='sh') and (path='/data' or path='/home')")
 	mockFilter.EXPECT().
 		FindSeriesIDsByExpr(uint32(10), &stmt.InExpr{Key: "ip", Values: []string{"1.1.1.1", "2.2.2.2"}}, query.TimeRange).
-		Return(mockSeriesIDSet(int64(11), roaring.BitmapOf(1, 2, 4)), nil)
+		Return(mockSeriesIDSet(uint32(11), roaring.BitmapOf(1, 2, 4)), nil)
 	mockFilter.EXPECT().
 		GetSeriesIDsForTag(uint32(10), "ip", query.TimeRange).
-		Return(mockSeriesIDSet(int64(11), roaring.BitmapOf(1, 2, 3, 4, 6, 7, 8)), nil)
+		Return(mockSeriesIDSet(uint32(11), roaring.BitmapOf(1, 2, 3, 4, 6, 7, 8)), nil)
 	mockFilter.EXPECT().
 		FindSeriesIDsByExpr(uint32(10), &stmt.EqualsExpr{Key: "region", Value: "sh"}, query.TimeRange).
-		Return(mockSeriesIDSet(int64(11), roaring.BitmapOf(2, 3, 4, 7)), nil)
+		Return(mockSeriesIDSet(uint32(11), roaring.BitmapOf(2, 3, 4, 7)), nil)
 	mockFilter.EXPECT().
 		FindSeriesIDsByExpr(uint32(10), &stmt.EqualsExpr{Key: "path", Value: "/data"}, query.TimeRange).
-		Return(mockSeriesIDSet(int64(11), roaring.BitmapOf(3, 5)), nil)
+		Return(mockSeriesIDSet(uint32(11), roaring.BitmapOf(3, 5)), nil)
 	mockFilter.EXPECT().
 		FindSeriesIDsByExpr(uint32(10), &stmt.EqualsExpr{Key: "path", Value: "/home"}, query.TimeRange).
-		Return(mockSeriesIDSet(int64(11), roaring.BitmapOf(1)), nil)
+		Return(mockSeriesIDSet(uint32(11), roaring.BitmapOf(1)), nil)
 	search := newSeriesSearch(10, mockFilter, query)
 	resultSet, _ := search.Search()
 	// ip not in ('1.1.1.1','2.2.2.2') => 3,6,7,8
 	// ip not in ('1.1.1.1','2.2.2.2') and region='sh' => 3,7
 	// path='/data' or path='/home' => 1,3,5
 	// final => 3
-	assert.Equal(t, *mockSeriesIDSet(int64(11), roaring.BitmapOf(3)), *resultSet)
+	assert.Equal(t, *mockSeriesIDSet(uint32(11), roaring.BitmapOf(3)), *resultSet)
 
 	// error
 	mockFilter1 := series.NewMockFilter(ctrl)
 	mockFilter1.EXPECT().
 		FindSeriesIDsByExpr(uint32(10), &stmt.InExpr{Key: "ip", Values: []string{"1.1.1.1", "2.2.2.2"}}, query.TimeRange).
-		Return(mockSeriesIDSet(int64(11), roaring.BitmapOf(1, 2, 4)), nil)
+		Return(mockSeriesIDSet(uint32(11), roaring.BitmapOf(1, 2, 4)), nil)
 	mockFilter1.EXPECT().
 		GetSeriesIDsForTag(uint32(10), "ip", query.TimeRange).
-		Return(mockSeriesIDSet(int64(11), roaring.BitmapOf(1, 2, 3, 4, 6, 7, 8)), nil)
+		Return(mockSeriesIDSet(uint32(11), roaring.BitmapOf(1, 2, 3, 4, 6, 7, 8)), nil)
 	mockFilter1.EXPECT().
 		FindSeriesIDsByExpr(uint32(10), &stmt.EqualsExpr{Key: "region", Value: "sh"}, query.TimeRange).
 		Return(nil, errors.New("complex error"))
@@ -203,7 +203,7 @@ func TestComplexCondition(t *testing.T) {
 	assert.NotNil(t, err)
 }
 
-func mockSeriesIDSet(version int64, ids *roaring.Bitmap) *series.MultiVerSeriesIDSet {
+func mockSeriesIDSet(version uint32, ids *roaring.Bitmap) *series.MultiVerSeriesIDSet {
 	s := series.NewMultiVerSeriesIDSet()
 	s.Add(version, ids)
 	return s
